Add a map entity for controlling the game
This doesn't currently do anything but I want to use it to hook into all
the various triggers dotted around the map for controlling things.

It also means that if this gamemode gets popular enough that people make
maps for it, they can just add all the triggers we have to retrofit into
existing maps for us and use hammer IO to talk to the gamemode.