Fix usage of "restart" meaning teleport to spawn
This was terrible naming on my part, probably caused by the way this
logic evolved when I implemented it.

Hopefully this mode is more understandable at first glance, and it also
opens up the ability to specify which spawn the player jumped into.

I didn't bother to implement that in this commit (because I'm lazy) but
at some point in the future.

Not sure what to do with that information but whatever it's there